Add README to concourse directory
This file explains the contract that we are drawing out for the use and maintenance of the concourse directory. It describes the organization of the directory, and how to make changes safely. It also describes individual team's responsibility to make sure they are alerted if changes are made to a pipeline they are using.

[#134241385]