explicitly state expected pipelines in README
It was brought up by Robert Eckhardt that team names don't mean much
with regards to OSS. This commit acknowledges that and replaces team
names with code specific justification for the excetions. It also
addresses Heikki's concern about being explicit with regards to which
pipelines should live in this directory.

[#134241385]